+++ src/features/store.ts
@@ -64,6 +64,7 @@ import { apiSlice as serviceAdminBoardApiSlice } from './adminBoard/serviceAdmin
 import { apiSlice as companyRoleApiSlice } from './companyRoles/companyRoleApiSlice'
 import { apiSlice as certificationApiSlice } from './certification/certificationApiSlice'
 import { apiSlice as userManagementApiSlice } from './appManagement/userManagementApiSlice'
+import { apiSlice as registrationApiSlice } from './registration/registrationApiSlice'
 
 import languageSlice from './language/slice'
 import { apiSlice as usecaseApiSlice } from './usecase/usecaseApiSlice'
@@ -117,6 +118,7 @@ export const reducers = {
   [certificationApiSlice.reducerPath]: certificationApiSlice.reducer,
   [userManagementApiSlice.reducerPath]: userManagementApiSlice.reducer,
   [usecaseApiSlice.reducerPath]: usecaseApiSlice.reducer,
+  [registrationApiSlice.reducerPath]: registrationApiSlice.reducer,
 }
 
 export const store = configureStore({
@@ -147,7 +149,8 @@ export const store = configureStore({
       .concat(companyRoleApiSlice.middleware)
       .concat(certificationApiSlice.middleware)
       .concat(userManagementApiSlice.middleware)
-      .concat(usecaseApiSlice.middleware),
+      .concat(usecaseApiSlice.middleware)
+      .concat(registrationApiSlice.middleware),
 })
 
 type RootState = ReturnType<typeof store.getState>